Update README.md
Reviewed and edited for clarity and IBM style. 

Any information in parenthesis, like in line 85 or 102, needs to be rewritten to be incorporated in the main text or removed because it's redundant or not needed. 

In line 70, before the in-depth explanation, it would be good to give just one sentence about what the resourceset component actually does, like

This feature allows you to obtain an instance of an object structure that holds a type of Maximo data.